Fix handling of slices in ConvertToPatchMap
The ConvertToPatchMap function in the common package was incorrectly
handling slices, causing a panic in certain scenarios. Specifically,
when converting a slice of strings, the code assumed a pointer to a
slice, leading to an interface conversion error.

This change addresses the issue by correctly checking the kind of the
reflect value and handling slices appropriately. Now, the function
correctly converts slices and avoids panics.

Test Plan:
PASS: Build the image and perform a first deploy.
PASS: Run a scenario where the panic occurs. Particularly, when
a PlatformNetwork is updated in a day-2 situation. Verify that the
panic is not happening.

Signed-off-by: Enzo Candotti <enzo.candotti@windriver.com>
(cherry picked from commit 26888044832daffa93f5d14a75478d368264a63e)